+++ src/seeds/users.ts
@@ -1,12 +1,13 @@
 import { type User } from '@prisma/client';
-import * as bcrypt from 'bcrypt';
+
+import { hashPassword } from '../utils/auth';
 
 const createUsersSeed = async (): Promise<User[]> => {
   const users = [
     {
       id: 1,
       username: 'Alex',
-      password: await bcrypt.hash('123456', 10),
+      password: await hashPassword('123456'),
       createdAt: new Date(),
       updatedAt: new Date(),
     },